requirements.txt
Django==2.1.7
mysqlclient==1.4.2.post1
pytz==2018.9
PyJWT==1.7.1
